Updated the C++ wrapper
- Changed the test routine wrappers to automatically use the _BraidApp**()
functions. This makes the test routine interfaces in C++ cleaner

- Changed the braid_Core wrapper to use braid_Int and braid_Real, not int
and double

- Changed the way the status structures are wrapped.  Now, the _BraidApp**()
routines do the conversion from the C structure "braid_**Status" to
the C++ object "Braid**Status".  This way the user routines don't have to do
this conversion.

- Moved the Coarsen and Refine routines into the App structure.  They work
the same way as all the other function Phi, Clone, and so on.  Except, these
functions are only virtual, meanining that their definition is optional.  The
default implementation just calls clone.

- Drive-04 has been changed to accommodate this update