Adding new option to skip all work on the down cycle.  This is now the default option in Braid.  It can be toggled, though, with the "braid_SetSkip" option.
I also cleaned up how the core->rnorms array is accessed with three new accessor
functions
  _braid_AppendResidual
  _braid_GetRNorm
  _braid_DeleteLasResidual
The number of residual norms available depends on the skip option and the
iteration number.  Sometimes there are resdiual norms available, sometimes not.
And the number available does not equal the number of iterations.  So, these
functions make all that easier.