First draft at implementing STMG in Braid
Options -stmg, -pfmg_mi, and -pfmg_ml were added to drive-02.
Options -pfmg_mi and -pfmg_ml were also added to ex-02.

Introduced an even/odd approach in FCRelax() when the coarsening factor is 1.
In this case, all points are C-points in time and Braid originally called Step()
in a Jacobi-like fashion.  We probably want to give the user the ability to tell
Braid what to do in this case (somehow) instead of hardcoding something.

Some renaming needs to happen for the spatial solver parameters to distinguish
them better from temporal grid parameters.  An attempt at this was made, but it
surely can be improved.