Added a way to store a better initial guess for implicit schemes.
When using the no-residual option, a better initial guess for Step() is the last
solution computed by Step().  This is a solution of the homogeneous problem, not
the FAS problem.  This version of the code stores this better initial guess in a
new array called wa[], which mimics the va[] array already in Braid.  It is only
used for the no-residual option.  The value is updated in two places: Step() and
Residual().  There are some subtleties related to the latter routine associated
with maintaining a fixed-point iteration that need to be documented.  Also, some
discussion is needed about how to control when/where to use this option.  One
idea is to give the user control over each grid level individually (in much the
same way as is done currently with the coarsening factor), since the additional
storage and copying is not needed on levels where explicit schemes are used.