Add shared pointer structure for the adjoint vectors.
Set up the braid_Adjoint structure that holds an braid_UserVector for
storing adjoint data as well as a useCount that counts the number of
pointer pointing to the structure.

UserInit and UserClone initialize adjoint vectors with zero and set
there useCount to 1. UserFree decreases the useCount and deletes the
adjoint if the useCount is zero.