Change name for adjoint function: Append '_diff' instead of 'adjoint'.
The idea is to use the name "adjoint" only for the adjoint variables of
the outer optimization problem, but not for the intermediate AD-based
variables during reverse tape evaluation. So far, only the names of the
function calls has been adapted by appending '_diff' instead of
'adjoint'.

TODO: Find another name for the shared pointer struct that holds the
primal as well as it's AD-'adjoint' counterpart. Maybe use '_bar'?