Richardson (#50)
Port and update of Ben ONeill's Richardson Extrapolation code.  This code adds a new feature that allows you to enable Richardson extrapolation (based on the first coarse MGRIT level) to enhance accuracy on the finest level, and to also get an error estimate on the finest level, e.g., for adaptive refinement.  See examples/ex-06.c  and  examples/ex-02.c and the function braid_SetRichardsonEstimation().